+++ src/v2019C1/extensions/identity/transaction/index.ts
@@ -1,2 +1,3 @@
 export { writeIdentity } from './write-identity'
+export { restoreClonedUuids } from './restore-cloned-uuids'
 export type { IdentityMode } from './write-identity.types'